bookmark; split side input storage manager and task instance
